adjust the chatbots as common functions
